Initial work on Spec validation
